[기능 수정] 크루 생성 컴포넌트 세분화, 파일 구조 변경
